http-apis/pyproject.toml
[tool.poetry]
name = "http-apis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
strawberry-graphql = {extras = ["debug-server", "fastapi"], version = "^0.260.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
